#6327 rework testRequestIdleTimeout and merge http(s) and fcgi test
